add the option to save generated images + make the config clearer

+++ gui_classes/gui_manager/thread_manager.py
@@ -9,6 +9,7 @@ from comfy_classes.comfy_class_API import ImageGeneratorAPIWrapper
 from gui_classes.gui_object.overlay import OverlayCountdown, OverlayLoading
 from gui_classes.gui_object.toolbox import ImageUtils
 from hotspot_classes.hotspot_client import HotspotClient
+from constant import KEEP_GENERATED_IMAGE
 
 import logging
 logger = logging.getLogger(__name__)
@@ -182,7 +183,7 @@ class CountdownThread(QObject):
 class ImageGenerationThread(QObject):
     finished = Signal(object)
 
-    def __init__(self, style: object, input_image: QImage, parent: QObject = None) -> None:
+    def __init__(self, style: object, input_image: QImage, api: ImageGeneratorAPIWrapper, parent: QObject = None) -> None:
         """
         Initialize the ImageGenerationThread with style, input image, and optional parent.
         """
@@ -191,7 +192,7 @@ class ImageGenerationThread(QObject):
         super().__init__(parent)
         self.style = style
         self.input_image = input_image
-        self.api = ImageGeneratorAPIWrapper(style=style, qimg=input_image)
+        self.api = api
         self._running = True
         self._thread = None
         self._worker = None
@@ -360,8 +361,12 @@ class ImageGenerationThread(QObject):
                         if DEBUG_ImageGenerationThread:
                             logger.info(f"[DEBUG][ImageGenerationWorker] Successfully loaded generated image.")
                         self.finished.emit(qimg)
-                        
-                    self.api.delete_input_and_output_images()
+                    
+                    self.image_path = self.api.get_latest_image_path()
+                    if not KEEP_GENERATED_IMAGE:
+                        self.api.delete_input_and_output_images()
+                    else:
+                        self.api.move_output_image()
                 except Exception as e:
                     if DEBUG_ImageGenerationThread:
                         logger.info(f"[DEBUG][ImageGenerationWorker] Exception: {e}")
